Add, update_attributes, remove_attributes, ansible check_mode

+++ plugins/modules/folder.py
@@ -32,7 +32,21 @@ options:
         type: str
         aliases: [title]
     attributes:
-        description: The attributes of your folder as described in the API documentation.
+        description:
+            - The attributes of your folder as described in the API documentation.
+              B(Attention! This option OVERWRITES all existing attributes!)
+        type: raw
+        default: {}
+    update_attributes:
+        description:
+            - The update_attributes of your host as described in the API documentation.
+              This will only update the given attributes.
+        type: raw
+        default: {}
+    remove_attributes:
+        description:
+            - The remove_attributes of your host as described in the API documentation.
+              This will only remove the given attributes.
         type: raw
         default: []
     state:
@@ -44,6 +58,7 @@ options:
 author:
     - Robin Gierse (@robin-tribe29)
     - Lars Getwan (@lgetwan)
+    - Michael Sekania (@msekania)
 """
 
 EXAMPLES = r"""
@@ -70,11 +85,48 @@ EXAMPLES = r"""
     attributes:
       site: "my_remote_site"
     state: "present"
+
+# Create a folder with Criticality set to a Test system and Networking Segment WAN (high latency)"
+- name: "Create a folder with tag_criticality test and tag_networking wan"
+  tribe29.checkmk.folder:
+    server_url: "http://localhost/"
+    site: "my_site"
+    automation_user: "automation"
+    automation_secret: "$SECRET"
+    path: "/my_remote_folder"
+    attributes:
+      tag_criticality: "test"
+      tag_networking: "wan"
+    state: "present"
+
+# Update only specified attributes
+- name: "Update only specified attributes"
+  tribe29.checkmk.folder:
+    server_url: "http://localhost/"
+    site: "my_site"
+    automation_user: "automation"
+    automation_secret: "$SECRET"
+    path: "/my_folder"
+    update_attributes:
+      tag_networking: "dmz"
+    state: "present"
+
+# Remove specified attributes
+- name: "Remove specified attributes"
+  tribe29.checkmk.folder:
+    server_url: "http://localhost/"
+    site: "my_site"
+    automation_user: "automation"
+    automation_secret: "$SECRET"
+    path: "/my_folder"
+    remove_attributes:
+      - tag_networking
+    state: "present"
 """
 
 RETURN = r"""
 message:
-    description: The output message that the module generates.
+    description: The output message that the module generates. Contains the API response details in case of an error.
     type: str
     returned: always
     sample: 'Folder created.'
@@ -83,9 +135,14 @@ message:
 import sys
 import traceback
 
+import json
+
 # https://docs.ansible.com/ansible/latest/dev_guide/testing/sanity/import.html
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
+from ansible.module_utils.common.dict_transformations import dict_merge
+from ansible.module_utils.common.dict_transformations import recursive_diff
 from ansible.module_utils.urls import fetch_url
+
 
 if sys.version[0] == "3":
     from pathlib import Path
@@ -101,9 +158,6 @@ else:
     else:
         HAS_PATHLIB2_LIBRARY = True
         PATHLIB2_LIBRARY_IMPORT_ERROR = None
-
-import json
-
 
 def exit_failed(module, msg):
     result = {"msg": msg, "changed": False, "failed": True}
@@ -126,16 +180,18 @@ def cleanup_path(path):
         p = Path("/").joinpath(p)
     return str(p.parent).lower(), p.name
 
+def path_for_url(module):
+    return module.params["path"].replace("/", "~")
 
 def get_current_folder_state(module, base_url, headers):
     current_state = "unknown"
     current_explicit_attributes = {}
+    current_title = ""
     etag = ""
 
-    path_for_url = module.params["path"].replace("/", "~")
-
-    api_endpoint = "/objects/folder_config/" + path_for_url
-    url = base_url + api_endpoint
+    api_endpoint = "/objects/folder_config/" + path_for_url(module)
+    parameters = "?show_hosts=false"
+    url = base_url + api_endpoint + parameters
 
     response, info = fetch_url(module, url, data=None, headers=headers, method="GET")
 
@@ -145,6 +201,7 @@ def get_current_folder_state(module, base_url, headers):
         etag = info.get("etag", "")
         extensions = body.get("extensions", {})
         current_explicit_attributes = extensions.get("attributes", {})
+        current_title = "%s" % body.get("title", "")
         if "meta_data" in current_explicit_attributes:
             del current_explicit_attributes["meta_data"]
 
@@ -154,22 +211,43 @@ def get_current_folder_state(module, base_url, headers):
     else:
         exit_failed(
             module,
-            "Error calling API. HTTP code %d. Details: %s"
+            "Error calling API. HTTP code %d. Details: %s."
             % (info["status"], info.get("body", "N/A")),
         )
 
-    return current_state, current_explicit_attributes, etag
+    return current_state, current_explicit_attributes, current_title, etag
 
 
-def set_folder_attributes(module, attributes, base_url, headers):
-    parent, foldername = cleanup_path(module.params["path"])
-    path_for_url = module.params["path"].replace("/", "~")
-    name = module.params.get("name", foldername)
-
-    api_endpoint = "/objects/folder_config/" + path_for_url
+def set_folder_attributes(module, attributes, base_url, headers, update_method="attributes"):
+    api_endpoint = "/objects/folder_config/" + path_for_url(module)
     params = {
-        "name": name,
-        "attributes": attributes,
+        update_method: attributes,
+    }
+    url = base_url + api_endpoint
+
+    response, info = fetch_url(
+        module, url, module.jsonify(params), headers=headers, method="PUT"
+    )
+
+    if info["status"] == 400 and update_method == "remove_attributes":
+        # "Folder attributes allready removed."
+        return False
+    elif info["status"] != 200:
+        exit_failed(
+            module,
+            "Error calling API. HTTP code %d. Details: %s, "
+            % (info["status"], info["body"]),
+        )
+
+    return True
+
+
+def change_title(module, attributes, base_url, headers):
+    name = module.params.get("name")
+
+    api_endpoint = "/objects/folder_config/" + path_for_url(module)
+    params = {
+        "title": name,
     }
     url = base_url + api_endpoint
 
@@ -183,7 +261,6 @@ def set_folder_attributes(module, attributes, base_url, headers):
             "Error calling API. HTTP code %d. Details: %s, "
             % (info["status"], info["body"]),
         )
-
 
 def create_folder(module, attributes, base_url, headers):
     parent, foldername = cleanup_path(module.params["path"])
@@ -211,9 +288,7 @@ def create_folder(module, attributes, base_url, headers):
 
 
 def delete_folder(module, base_url, headers):
-    path_for_url = module.params["path"].replace("/", "~")
-
-    api_endpoint = "/objects/folder_config/" + path_for_url
+    api_endpoint = "/objects/folder_config/" + path_for_url(module)
     url = base_url + api_endpoint
 
     response, info = fetch_url(module, url, data=None, headers=headers, method="DELETE")
@@ -239,11 +314,13 @@ def run_module():
             required=False,
             aliases=["title"],
         ),
-        attributes=dict(type="raw", default=[]),
+        attributes=dict(type="raw", default={}),
+        remove_attributes=dict(type="raw", default=[]),
+        update_attributes=dict(type="raw", default={}),
         state=dict(type="str", default="present", choices=["present", "absent"]),
     )
 
-    module = AnsibleModule(argument_spec=module_args, supports_check_mode=False)
+    module = AnsibleModule(argument_spec=module_args, supports_check_mode=True)
 
     # Handle library import error according to the following link:
     # https://docs.ansible.com/ansible/latest/dev_guide/testing/sanity/import.html
@@ -272,6 +349,8 @@ def run_module():
 
     # Determine desired state and attributes
     attributes = module.params.get("attributes", {})
+    remove_attributes = module.params.get("remove_attributes", [])
+    update_attributes = module.params.get("update_attributes", {})
     if attributes == []:
         attributes = {}
     state = module.params.get("state", "present")
@@ -280,6 +359,7 @@ def run_module():
     (
         current_state,
         current_explicit_attributes,
+        current_title,
         etag,
     ) = get_current_folder_state(module, base_url, headers)
 
@@ -288,9 +368,43 @@ def run_module():
         headers["If-Match"] = etag
         msg_tokens = []
 
+        merged_attributes = dict_merge(current_explicit_attributes, update_attributes)
+
+        if module.params["name"] and current_title != module.params["name"]:
+            if not module.check_mode:
+                change_title(module, attributes, base_url, headers)
+            msg_tokens.append("Folder title updated.")
+
         if attributes != {} and current_explicit_attributes != attributes:
-            set_folder_attributes(module, attributes, base_url, headers)
-            msg_tokens.append("Folder attributes changed.")
+            if not module.check_mode:
+                set_folder_attributes(module, attributes, base_url, headers, "attributes")
+            msg_tokens.append("Folder attributes replaced.")
+
+        if update_attributes != {} and current_explicit_attributes != dict_merge(current_explicit_attributes, merged_attributes):
+            if not module.check_mode:
+                set_folder_attributes(
+                    module, merged_attributes, base_url, headers, "update_attributes"
+                )
+            msg_tokens.append("Folder attributes updated.")
+
+        if remove_attributes != []:
+            attr_exists = False
+            for el in remove_attributes:
+                if current_explicit_attributes.get(el):
+                    attr_exists = True
+                    break
+            changed = attr_exists
+            if not module.check_mode:
+                changed = set_folder_attributes(
+                    module, remove_attributes, base_url, headers, "remove_attributes"
+                )
+
+            if changed:
+                msg_tokens.append("Folder attributes removed.")
+            elif len(msg_tokens) >= 1:
+                msg_tokens.append("Folder attributes allready removed.")
+            else:
+                exit_ok(module, "Folder attributes allready removed.")
 
         if len(msg_tokens) >= 1:
             exit_changed(module, " ".join(msg_tokens))
@@ -300,14 +414,20 @@ def run_module():
             )
 
     elif state == "present" and current_state == "absent":
-        create_folder(module, attributes, base_url, headers)
+        if update_attributes != {} and attributes == {}:
+            attributes = update_attributes
+        if not module.params["name"]:
+            module.params["name"] = ""
+        if not module.check_mode:
+            create_folder(module, attributes, base_url, headers)
         exit_changed(module, "Folder created.")
 
     elif state == "absent" and current_state == "absent":
         exit_ok(module, "Folder already absent.")
 
     elif state == "absent" and current_state == "present":
-        delete_folder(module, base_url, headers)
+        if not module.check_mode:
+            delete_folder(module, base_url, headers)
         exit_changed(module, "Folder deleted.")
 
     else:
